Update and rename direct_control.m to pc/direct_control.m

+++ pc/direct_control.m
@@ -10,7 +10,7 @@ srl_flush(s1);
 
 while 1
 ipt=input('Enter command:' );
-ipt=uint16(ipt);
+ipt=uint8(ipt);
 disp(ipt);
 srl_write(s1,ipt);
 opt=srl_read(s1,10);
